refactor: keep initialization of service in init method
